sql: Create a builtin to reset the zone configs of a multi-region table
The new(ish) multi-region syntax introduced in 21.1 sets zone
configurations at the table or partition level for most of the table
localities. These zone configurations can diverge from their original
state either because the user has manually overridden them, or because
the system hasn't kept them up to date as the table has evolved (e.g.
with issue #66637). In either case, the user may wish to revert the zone
configurations for a given table back to their original state.

The new builtin takes a table ID, and resets its underlying zone
configurations to their original state (as specified by the assigned
table locality). For non-multi-region tables, the builtin no-ops.

Release note (sql change): Creates a builtin to reset the zone
configurations of multi-region tables. This builtin can be helpful in
cases where the user has overridden the zone configuration for a given
table and wishes to revert back to the original system-specified state.