amr: First cut of sampling
Implements sampling syntax and first cut of sampling for reads _only_. Works
for simple testing, but more testing is needed. Also still need to add sampling
for writes.

Also requires considerable code cleanup to encapsulate the sampling functions
(especially the SHA generated sampling, which has to be identical for reads and
writes for the tracking to work at all).

Release note: None