multitenant: Increase retry for upgrade rpc
The multi-tenant upgrade interlock requires that all SQL servers be contacted
during an upgrade. In cases where a SQL server is stopped and restarted shortly
before (or during) an upgrade, the RPC retry may fail before the connection to
the stopped/started SQL server is reestablished. This commit increases the
duration of the retry to 1 second to ensure that we don't fail prematurely.

Fixes: #114004
Release note: None